moved navigation.ts to src/router
